chore(client): adapt to unique partition ids
- remove topic name from messages where it is not required anymore
- temporary change for task/topic subscriptions: allow users to specify
  a partition id when creating a subscription; if no id is specified an
  arbitrary partition id is chosen (use case: single partition)

related to #414

+++ broker-core/src/main/java/io/zeebe/broker/logstreams/LogStreamsManager.java
@@ -55,7 +55,7 @@ public class LogStreamsManager
     }
 
 
-    public LogStream getLogStream(final long partitionId)
+    public LogStream getLogStream(final int partitionId)
     {
         return logStreams.get(partitionId);
     }
